task_1/requirements.txt
ImageHash==4.3.1
matplotlib==3.6.1
pandas==1.5.0
Pillow==9.2.0
scikit_learn==1.1.2
